app/requirements.txt
coloredlogs==15.0.1
pika==1.3.1
pydub==0.25.1
python-dotenv==0.21.1
redis==4.5.1
requests==2.28.2
sentry_sdk==1.15.0
sox==1.4.1
watchdog==2.2.1
